+++ plone/app/tiles/__init__.py
@@ -1,5 +1,19 @@
 # -*- coding: utf-8 -*-
-import zope.i18nmessageid
+from zope.i18nmessageid import MessageFactory
+from Products.CMFCore.permissions import ManagePortal
+
+_ = MessageFactory = MessageFactory('plone')
 
 
-MessageFactory = zope.i18nmessageid.MessageFactory('plone')
+def initialize(context):
+    """Registers modifiers with zope (on zope startup).
+    """
+    from plone.app.tiles.modifiers import modifiers
+
+    for modifier in modifiers:
+        context.registerClass(
+            modifier['wrapper'], modifier['id'],
+            permission=ManagePortal,
+            constructors=(modifier['form'], modifier['factory']),
+            icon=modifier['icon'],
+        )